making master branch an old commit

+++ BackEnd/src/main/java/model/Availability.java
@@ -5,11 +5,11 @@ public class Availability {
     private int eventId;
     private int personId;
     private int availabilityId;
-    private double startTime;
-    private double endTime;
+    private int startTime;
+    private int endTime;
     private int dow;
 
-    public Availability(int eId, int pId, double st, double et, int dow) {
+    public Availability(int eId, int pId, int st, int et, int dow) {
         this.eventId=eId;
         this.personId=pId;
         this.startTime=st;
@@ -41,19 +41,19 @@ public class Availability {
         this.availabilityId = availabilityId;
     }
 
-    public double getStartTime() {
+    public int getStartTime() {
         return startTime;
     }
 
-    public void setStartTime(double startTime) {
+    public void setStartTime(int startTime) {
         this.startTime = startTime;
     }
 
-    public double getEndTime() {
+    public int getEndTime() {
         return endTime;
     }
 
-    public void setEndTime(double endTime) {
+    public void setEndTime(int endTime) {
         this.endTime = endTime;
     }
