+++ FrontEnd/src/main/java/dao/EventDao.java
@@ -1,6 +1,7 @@
 package dao;
 
 import exception.DaoException;
+import model.Availability;
 import model.Event;
 
 import java.util.List;
@@ -14,5 +15,5 @@ public interface EventDao {
 
     List<Event> findEventbyId(int id);
 
-    void removeAndUpdateOptTime(int eventId, double optTime);
+    void removeAndUpdateOptTime(int eventId, int optTime);
 }
